Move the NotDeletedQuerySet to flask_common.mongo.querysets
This is a more appropriate place for it.
