+++ flask_common/mongo/documents.py
@@ -7,12 +7,13 @@ from mongoengine import (
     DateTimeField,
     Document,
     OperationError,
-    Q,
     QuerySet,
     StringField,
     ValidationError,
     queryset_manager,
 )
+
+from .querysets import NotDeletedQuerySet
 
 
 class StringIdField(StringField):
@@ -104,29 +105,6 @@ class DocumentBase(Document):
         if update_date and 'set__date_updated' not in kwargs:
             kwargs['set__date_updated'] = datetime.datetime.utcnow()
         super(DocumentBase, self).update(*args, **kwargs)
-
-
-class NotDeletedQuerySet(QuerySet):
-    def __call__(
-        self,
-        q_obj=None,
-        class_check=True,
-        slave_okay=False,
-        read_preference=None,
-        **query
-    ):
-        # we don't use __ne=True here, because $ne isn't a selective query and doesn't utilize an index in the most efficient manner (http://docs.mongodb.org/manual/faq/indexes/#using-ne-and-nin-in-a-query-is-slow-why)
-        extra_q_obj = Q(is_deleted=False)
-        q_obj = q_obj & extra_q_obj if q_obj else extra_q_obj
-        return super(NotDeletedQuerySet, self).__call__(
-            q_obj, class_check, slave_okay, read_preference, **query
-        )
-
-    def count(self, *args, **kwargs):
-        # we need this hack for doc.objects.count() to exclude deleted objects
-        if not getattr(self, '_not_deleted_query_applied', False):
-            self = self.all()
-        return super(NotDeletedQuerySet, self).count(*args, **kwargs)
 
 
 class SoftDeleteDocument(Document):
